Make it consistent with https://github.com/DataDog/datadog-agent/pull/20372
