+++ comp/process/connectionscheck/connectionscheckimpl/check.go
@@ -3,14 +3,15 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package impl implements a component to handle realtime Container data collection in the Process Agent.
-package impl
+// Package connectionscheckimpl implements a component to handle Connections data collection in the Process Agent.
+package connectionscheckimpl
 
 import (
 	"go.uber.org/fx"
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
-	"github.com/DataDog/datadog-agent/comp/process/rtcontainercheck"
+	"github.com/DataDog/datadog-agent/comp/core/sysprobeconfig"
+	"github.com/DataDog/datadog-agent/comp/process/connectionscheck"
 	"github.com/DataDog/datadog-agent/comp/process/types"
 	"github.com/DataDog/datadog-agent/pkg/process/checks"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
@@ -24,25 +25,26 @@ var Module = fxutil.Component(
 var _ types.CheckComponent = (*check)(nil)
 
 type check struct {
-	rtContainerCheck *checks.RTContainerCheck
+	connectionsCheck *checks.ConnectionsCheck
 }
 
 type dependencies struct {
 	fx.In
 
-	Config config.Component
+	Sysconfig sysprobeconfig.Component
+	Config    config.Component
 }
 
 type result struct {
 	fx.Out
 
 	Check     types.ProvidesCheck
-	Component rtcontainercheck.Component
+	Component connectionscheck.Component
 }
 
 func newCheck(deps dependencies) result {
 	c := &check{
-		rtContainerCheck: checks.NewRTContainerCheck(deps.Config),
+		connectionsCheck: checks.NewConnectionsCheck(deps.Config, deps.Sysconfig, deps.Sysconfig.SysProbeObject()),
 	}
 	return result{
 		Check: types.ProvidesCheck{
@@ -53,5 +55,5 @@ func newCheck(deps dependencies) result {
 }
 
 func (c *check) Object() checks.Check {
-	return c.rtContainerCheck
+	return c.connectionsCheck
 }
