+++ comp/process/containercheck/containercheckimpl/mock.go
@@ -5,7 +5,7 @@
 
 //go:build test
 
-package impl
+package containercheckimpl
 
 import (
 	"testing"
