+++ comp/process/expvars/expvarsimpl/expvars_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package impl
+package expvarsimpl
 
 import (
 	"net/http"
@@ -16,7 +16,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core"
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/comp/process/expvars"
-	hostinfo "github.com/DataDog/datadog-agent/comp/process/hostinfo/impl"
+	"github.com/DataDog/datadog-agent/comp/process/hostinfo/hostinfoimpl"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
@@ -25,7 +25,7 @@ func TestExpvarServer(t *testing.T) {
 		fx.Supply(core.BundleParams{}),
 
 		Module,
-		hostinfo.MockModule,
+		hostinfoimpl.MockModule,
 		core.MockBundle,
 	))
 
@@ -48,7 +48,7 @@ func TestTelemetry(t *testing.T) {
 		}}),
 
 		Module,
-		hostinfo.MockModule,
+		hostinfoimpl.MockModule,
 		core.MockBundle,
 	))
