+++ comp/process/hostinfo/hostinfoimpl/hostinfo.go
@@ -3,8 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package impl wraps the hostinfo inside a component. This is useful because it is relied on by other components.
-package impl
+// Package hostinfoimpl wraps the hostinfo inside a component. This is useful because it is relied on by other components.
+package hostinfoimpl
 
 import (
 	"fmt"
