--- comp/process/hostinfo/impl/hostinfo_mock.go
@@ -1,25 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-//go:build test
-
-package impl
-
-import (
-	"go.uber.org/fx"
-
-	hostinfoComp "github.com/DataDog/datadog-agent/comp/process/hostinfo"
-	"github.com/DataDog/datadog-agent/pkg/process/checks"
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
-)
-
-// MockModule defines the fx options for the mock component.
-var MockModule = fxutil.Component(
-	fx.Provide(newMockHostInfo),
-)
-
-func newMockHostInfo() hostinfoComp.Component {
-	return &hostinfo{hostinfo: &checks.HostInfo{}}
-}
