+++ comp/process/profiler/profilerimpl/profiler.go
@@ -3,8 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package impl implements a component to handle starting and stopping the internal profiler.
-package impl
+// Package profilerimpl implements a component to handle starting and stopping the internal profiler.
+package profilerimpl
 
 import (
 	"context"
