+++ comp/process/runner/runnerimpl/runner.go
@@ -3,8 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package impl implements a component to run data collection checks in the Process Agent.
-package impl
+// Package runnerimpl implements a component to run data collection checks in the Process Agent.
+package runnerimpl
 
 import (
 	"context"
@@ -14,7 +14,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/comp/core/sysprobeconfig"
 	"github.com/DataDog/datadog-agent/comp/process/hostinfo"
-	runnerComp "github.com/DataDog/datadog-agent/comp/process/runner"
+	"github.com/DataDog/datadog-agent/comp/process/runner"
 	"github.com/DataDog/datadog-agent/comp/process/submitter"
 	"github.com/DataDog/datadog-agent/comp/process/types"
 	"github.com/DataDog/datadog-agent/pkg/process/checks"
@@ -28,7 +28,7 @@ var Module = fxutil.Component(
 )
 
 // runner implements the Component.
-type runner struct {
+type runnerImpl struct {
 	checkRunner    *processRunner.CheckRunner
 	providedChecks []types.CheckComponent
 }
@@ -46,14 +46,14 @@ type dependencies struct {
 	Config   config.Component
 }
 
-func newRunner(deps dependencies) (runnerComp.Component, error) {
+func newRunner(deps dependencies) (runner.Component, error) {
 	c, err := processRunner.NewRunner(deps.Config, deps.SysCfg.SysProbeObject(), deps.HostInfo.Object(), filterEnabledChecks(deps.Checks), deps.RTNotifier)
 	if err != nil {
 		return nil, err
 	}
 	c.Submitter = deps.Submitter
 
-	runner := &runner{
+	runner := &runnerImpl{
 		checkRunner:    c,
 		providedChecks: deps.Checks,
 	}
@@ -66,11 +66,11 @@ func newRunner(deps dependencies) (runnerComp.Component, error) {
 	return runner, nil
 }
 
-func (r *runner) Run(context.Context) error {
+func (r *runnerImpl) Run(context.Context) error {
 	return r.checkRunner.Run()
 }
 
-func (r *runner) Stop(context.Context) error {
+func (r *runnerImpl) Stop(context.Context) error {
 	r.checkRunner.Stop()
 	return nil
 }
@@ -87,16 +87,16 @@ func filterEnabledChecks(providedChecks []types.CheckComponent) []checks.Check {
 
 // IsRealtimeEnabled checks the runner to see if it is running the process check in realtime mode.
 // This is primarily used in tests.
-func (r *runner) IsRealtimeEnabled() bool {
+func (r *runnerImpl) IsRealtimeEnabled() bool {
 	return r.checkRunner.IsRealTimeEnabled()
 }
 
 // GetChecks returns the checks that are currently enabled and provided to the runner
-func (r *runner) GetChecks() []checks.Check {
+func (r *runnerImpl) GetChecks() []checks.Check {
 	return r.checkRunner.GetChecks()
 }
 
 // GetProvidedChecks returns all provided checks, enabled or not.
-func (r *runner) GetProvidedChecks() []types.CheckComponent {
+func (r *runnerImpl) GetProvidedChecks() []types.CheckComponent {
 	return r.providedChecks
 }
